requirements.txt
langchain
langchain-community
chromadb
sentence-transformers
streamlit
python-dotenv